requirements.txt
Flask==2.3.3
twilio==8.13.0
openai==0.28.0
python-dotenv==1.0.0
gunicorn==21.2.0